删除应用是也删除所有管理员与该应该的关联

+++ configcenter-web/src/main/java/org/antframework/configcenter/web/controller/manage/AppManageConteoller.java
@@ -8,8 +8,9 @@
  */
 package org.antframework.configcenter.web.controller.manage;
 
+import org.antframework.boot.bekit.AntBekitException;
 import org.antframework.common.util.facade.AbstractQueryResult;
-import org.antframework.configcenter.facade.api.ConfigService;
+import org.antframework.common.util.facade.Status;
 import org.antframework.configcenter.facade.api.manage.AppManageService;
 import org.antframework.configcenter.facade.info.AppInfo;
 import org.antframework.configcenter.facade.order.manage.AddOrModifyAppOrder;
@@ -18,6 +19,8 @@ import org.antframework.configcenter.facade.order.manage.QueryAppOrder;
 import org.antframework.configcenter.facade.result.manage.AddOrModifyAppResult;
 import org.antframework.configcenter.facade.result.manage.DeleteAppResult;
 import org.antframework.configcenter.web.manager.facade.api.ManagerAppManageService;
+import org.antframework.configcenter.web.manager.facade.order.DeleteManagerAppByAppOrder;
+import org.antframework.configcenter.web.manager.facade.result.DeleteManagerAppByAppResult;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -32,8 +35,6 @@ public class AppManageConteoller extends AbstractController {
     private AppManageService appManageService;
     @Autowired
     private ManagerAppManageService managerAppManageService;
-    @Autowired
-    private ConfigService configService;
 
     /**
      * 添加或修改应用
@@ -59,9 +60,16 @@ public class AppManageConteoller extends AbstractController {
     @RequestMapping("/deleteApp")
     public DeleteAppResult deleteApp(String appCode) {
         assertAdmin();
+        // 先删除管理员和应用关联
+        DeleteManagerAppByAppOrder byAppOrder = new DeleteManagerAppByAppOrder();
+        byAppOrder.setAppCode(appCode);
+        DeleteManagerAppByAppResult byAppResult = managerAppManageService.deleteManagerAppByApp(byAppOrder);
+        if (!byAppResult.isSuccess()) {
+            throw new AntBekitException(Status.FAIL, byAppResult.getCode(), byAppResult.getMessage());
+        }
+        // 删除应用
         DeleteAppOrder order = new DeleteAppOrder();
         order.setAppCode(appCode);
-
         return appManageService.deleteApp(order);
     }
 
@@ -81,54 +89,4 @@ public class AppManageConteoller extends AbstractController {
         order.setAppCode(appCode);
         return appManageService.queryApp(order);
     }
-
-//    // 超级管理员查询应用
-//    private QueryAppResult queryAppByAdmin(int pageNo, int pageSize, String appCode) {
-//        QueryAppOrder order = new QueryAppOrder();
-//        order.setPageNo(pageNo);
-//        order.setPageSize(pageSize);
-//        order.setAppCode(appCode);
-//        return appManageService.queryApp(order);
-//    }
-//
-//    // 普通管理员查询应用
-//    private AbstractQueryResult<AppInfo> queryAppByNormalManager(int pageNo, int pageSize, String appCode) {
-//        QueryManagedAppOrder order = new QueryManagedAppOrder();
-//        order.setPageNo(pageNo);
-//        order.setPageSize(pageSize);
-//        order.setManagerCode(SessionAccessor.getManagerInfo().getManagerCode());
-//        order.setAppCode(appCode);
-//        QueryManagedAppResult queryManagedAppResult = managerAppManageService.queryManagedApp(order);
-//        if (!queryManagedAppResult.isSuccess()) {
-//            throw new AntBekitException(Status.FAIL, queryManagedAppResult.getCode(), queryManagedAppResult.getMessage());
-//        }
-//
-//        AbstractQueryResult<AppInfo> result = new AbstractQueryResult<AppInfo>() {
-//        };
-//        result.setStatus(Status.SUCCESS);
-//        result.setCode(CommonResultCode.SUCCESS.getCode());
-//        result.setMessage(CommonResultCode.SUCCESS.getMessage());
-//        result.setTotalCount(queryManagedAppResult.getTotalCount());
-//        for (AppInfo appInfo : getAppInfos(queryManagedAppResult.getInfos())) {
-//            result.addInfo(appInfo);
-//        }
-//
-//        return result;
-//    }
-//
-//    // 获取应用信息
-//    private List<AppInfo> getAppInfos(List<ManagerAppInfo> managerAppInfos) {
-//        List<AppInfo> appInfos = new ArrayList<>();
-//        for (ManagerAppInfo info : managerAppInfos) {
-//            FindAppOrder order = new FindAppOrder();
-//            order.setAppCode(info.getAppCode());
-//            FindAppResult result = configService.findApp(order);
-//            if (!result.isSuccess()) {
-//                throw new AntBekitException(Status.FAIL, result.getCode(), result.getMessage());
-//            }
-//            appInfos.add(result.getAppInfo());
-//        }
-//
-//        return appInfos;
-//    }
 }
